backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pymupdf>=1.24.0
langchain>=0.3.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0
faiss-cpu>=1.9.0
numpy>=1.26.0
pydantic>=2.9.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0

